+++ src/ext/SimplexNoise.cc
@@ -95,11 +95,11 @@ static const int simplex[64][4] = {
   {2,1,0,3},{0,0,0,0},{0,0,0,0},{0,0,0,0},{3,1,0,2},{0,0,0,0},{3,2,0,1},{3,2,1,0}
 };
 
-int fastfloor(double x) { return x > 0 ? (int) x : (int) x - 1; }
+int fastfloor(float x) { return x > 0 ? (int) x : (int) x - 1; }
 
-double dot(int* g, double x, double y)                     { return g[0]*x + g[1]*y; }
-double dot(int* g, double x, double y, double z)           { return g[0]*x + g[1]*y + g[2]*z; }
-double dot(int* g, double x, double y, double z, double w) { return g[0]*x + g[1]*y + g[2]*z + g[3]*w; }
+float dot(int* g, float x, float y)                     { return g[0]*x + g[1]*y; }
+float dot(int* g, float x, float y, float z)           { return g[0]*x + g[1]*y + g[2]*z; }
+float dot(int* g, float x, float y, float z, float w) { return g[0]*x + g[1]*y + g[2]*z + g[3]*w; }
 
 
 // 2D Multi-octave Simplex noise.
@@ -107,8 +107,8 @@ double dot(int* g, double x, double y, double z, double w) { return g[0]*x + g[1
 // For each octave, a higher frequency/lower amplitude function will be added to
 // the original. The higher the persistence [0-1], the more of each succeeding
 // octave will be added.
-double layered(double octaves, double persistence, double x, double y) {
-  double total = 0, freq = 1, amplitude = 1, maxAmplitude = 0;
+float layered(float octaves, float persistence, float x, float y) {
+  float total = 0, freq = 1, amplitude = 1, maxAmplitude = 0;
 
   for (int i = 0; i < octaves; i++) {
     total += noise(x * freq, y * freq) * amplitude;
@@ -123,8 +123,8 @@ double layered(double octaves, double persistence, double x, double y) {
 // For each octave, a higher freq/lower amplitude function will be added to
 // the original. The higher the persistence [0-1], the more of each succeeding
 // octave will be added.
-double layered(double octaves, double persistence, double x, double y, double z) {
-  double total = 0, freq = 1, amplitude = 1, maxAmplitude = 0;
+float layered(float octaves, float persistence, float x, float y, float z) {
+  float total = 0, freq = 1, amplitude = 1, maxAmplitude = 0;
   
   for (int i = 0; i < octaves; i++) {
     total += noise(x * freq, y * freq, z * freq) * amplitude;
@@ -139,9 +139,9 @@ double layered(double octaves, double persistence, double x, double y, double z)
 // For each octave, a higher freq/lower amplitude function will be added to
 // the original. The higher the persistence [0-1], the more of each succeeding
 // octave will be added.
-double layered(double octaves, double persistence, 
-               double x, double y, double z, double w) {
-  double total = 0, freq = 1, amplitude = 1, maxAmplitude = 0;
+float layered(float octaves, float persistence, 
+               float x, float y, float z, float w) {
+  float total = 0, freq = 1, amplitude = 1, maxAmplitude = 0;
   
   for (int i = 0; i < octaves; i++) {
     total += noise(x * freq, y * freq, z * freq, w * freq) * amplitude;
@@ -152,30 +152,30 @@ double layered(double octaves, double persistence,
 }
 
 // 1D raw Simplex noise 
-double noise(double x) {
+float noise(float x) {
   return noise(x, 0);
 }
 
 // 2D raw Simplex noise
-double noise(double x, double y) {
+float noise(float x, float y) {
   // Noise contributions from the three corners
-  double n0, n1, n2;
+  float n0, n1, n2;
 
   // Skew the input space to determine which simplex cell we're in
-  double F2 = 0.5 * (sqrtf(3.0) - 1.0);
+  float F2 = 0.5 * (sqrtf(3.0) - 1.0);
   // Hairy factor for 2D
-  double s = (x + y) * F2;
+  float s = (x + y) * F2;
   int i = fastfloor( x + s );
   int j = fastfloor( y + s );
 
-  double G2 = (3.0 - sqrtf(3.0)) / 6.0;
-  double t = (i + j) * G2;
+  float G2 = (3.0 - sqrtf(3.0)) / 6.0;
+  float t = (i + j) * G2;
   // Unskew the cell origin back to (x,y) space
-  double X0 = i-t;
-  double Y0 = j-t;
+  float X0 = i-t;
+  float Y0 = j-t;
   // The x,y distances from the cell origin
-  double x0 = x-X0;
-  double y0 = y-Y0;
+  float x0 = x-X0;
+  float y0 = y-Y0;
 
   // For the 2D case, the simplex shape is an equilateral triangle.
   // Determine which simplex we are in.
@@ -186,10 +186,10 @@ double noise(double x, double y) {
   // A step of (1,0) in (i,j) means a step of (1-c,-c) in (x,y), and
   // a step of (0,1) in (i,j) means a step of (-c,1-c) in (x,y), where
   // c = (3-sqrt(3))/6
-  double x1 = x0 - i1 + G2; // Offsets for middle corner in (x,y) unskewed coords
-  double y1 = y0 - j1 + G2;
-  double x2 = x0 - 1.0 + 2.0 * G2; // Offsets for last corner in (x,y) unskewed coords
-  double y2 = y0 - 1.0 + 2.0 * G2;
+  float x1 = x0 - i1 + G2; // Offsets for middle corner in (x,y) unskewed coords
+  float y1 = y0 - j1 + G2;
+  float x2 = x0 - 1.0 + 2.0 * G2; // Offsets for last corner in (x,y) unskewed coords
+  float y2 = y0 - 1.0 + 2.0 * G2;
 
   // Work out the hashed gradient indices of the three simplex corners
   int ii = i & 255;
@@ -199,21 +199,21 @@ double noise(double x, double y) {
   int gi2 = perm[ii+1+perm[jj+1]] % 12;
 
   // Calculate the contribution from the three corners
-  double t0 = 0.5 - x0*x0-y0*y0;
+  float t0 = 0.5 - x0*x0-y0*y0;
   if(t0<0) n0 = 0.0;
   else {
       t0 *= t0;
       n0 = t0 * t0 * dot(grad3[gi0], x0, y0); // (x,y) of grad3 used for 2D gradient
   }
 
-  double t1 = 0.5 - x1*x1-y1*y1;
+  float t1 = 0.5 - x1*x1-y1*y1;
   if(t1<0) n1 = 0.0;
   else {
       t1 *= t1;
       n1 = t1 * t1 * dot(grad3[gi1], x1, y1);
   }
 
-  double t2 = 0.5 - x2*x2-y2*y2;
+  float t2 = 0.5 - x2*x2-y2*y2;
   if(t2<0) n2 = 0.0;
   else {
       t2 *= t2;
@@ -227,24 +227,24 @@ double noise(double x, double y) {
 
 
 // 3D raw Simplex noise
-double noise( double x, double y, double z ) {
-  double n0, n1, n2, n3; // Noise contributions from the four corners
+float noise( float x, float y, float z ) {
+  float n0, n1, n2, n3; // Noise contributions from the four corners
 
   // Skew the input space to determine which simplex cell we're in
-  double F3 = 1.0/3.0;
-  double s = (x+y+z)*F3; // Very nice and simple skew factor for 3D
+  float F3 = 1.0/3.0;
+  float s = (x+y+z)*F3; // Very nice and simple skew factor for 3D
   int i = fastfloor(x+s);
   int j = fastfloor(y+s);
   int k = fastfloor(z+s);
 
-  double G3 = 1.0/6.0; // Very nice and simple unskew factor, too
-  double t = (i+j+k)*G3;
-  double X0 = i-t; // Unskew the cell origin back to (x,y,z) space
-  double Y0 = j-t;
-  double Z0 = k-t;
-  double x0 = x-X0; // The x,y,z distances from the cell origin
-  double y0 = y-Y0;
-  double z0 = z-Z0;
+  float G3 = 1.0/6.0; // Very nice and simple unskew factor, too
+  float t = (i+j+k)*G3;
+  float X0 = i-t; // Unskew the cell origin back to (x,y,z) space
+  float Y0 = j-t;
+  float Z0 = k-t;
+  float x0 = x-X0; // The x,y,z distances from the cell origin
+  float y0 = y-Y0;
+  float z0 = z-Z0;
 
   // For the 3D case, the simplex shape is a slightly irregular tetrahedron.
   // Determine which simplex we are in.
@@ -266,15 +266,15 @@ double noise( double x, double y, double z ) {
   // a step of (0,1,0) in (i,j,k) means a step of (-c,1-c,-c) in (x,y,z), and
   // a step of (0,0,1) in (i,j,k) means a step of (-c,-c,1-c) in (x,y,z), where
   // c = 1/6.
-  double x1 = x0 - i1 + G3; // Offsets for second corner in (x,y,z) coords
-  double y1 = y0 - j1 + G3;
-  double z1 = z0 - k1 + G3;
-  double x2 = x0 - i2 + 2.0*G3; // Offsets for third corner in (x,y,z) coords
-  double y2 = y0 - j2 + 2.0*G3;
-  double z2 = z0 - k2 + 2.0*G3;
-  double x3 = x0 - 1.0 + 3.0*G3; // Offsets for last corner in (x,y,z) coords
-  double y3 = y0 - 1.0 + 3.0*G3;
-  double z3 = z0 - 1.0 + 3.0*G3;
+  float x1 = x0 - i1 + G3; // Offsets for second corner in (x,y,z) coords
+  float y1 = y0 - j1 + G3;
+  float z1 = z0 - k1 + G3;
+  float x2 = x0 - i2 + 2.0*G3; // Offsets for third corner in (x,y,z) coords
+  float y2 = y0 - j2 + 2.0*G3;
+  float z2 = z0 - k2 + 2.0*G3;
+  float x3 = x0 - 1.0 + 3.0*G3; // Offsets for last corner in (x,y,z) coords
+  float y3 = y0 - 1.0 + 3.0*G3;
+  float z3 = z0 - 1.0 + 3.0*G3;
 
   // Work out the hashed gradient indices of the four simplex corners
   int ii = i & 255;
@@ -286,28 +286,28 @@ double noise( double x, double y, double z ) {
   int gi3 = perm[ii+1+perm[jj+1+perm[kk+1]]] % 12;
 
   // Calculate the contribution from the four corners
-  double t0 = 0.6 - x0*x0 - y0*y0 - z0*z0;
+  float t0 = 0.6 - x0*x0 - y0*y0 - z0*z0;
   if(t0<0) n0 = 0.0;
   else {
       t0 *= t0;
       n0 = t0 * t0 * dot(grad3[gi0], x0, y0, z0);
   }
 
-  double t1 = 0.6 - x1*x1 - y1*y1 - z1*z1;
+  float t1 = 0.6 - x1*x1 - y1*y1 - z1*z1;
   if(t1<0) n1 = 0.0;
   else {
       t1 *= t1;
       n1 = t1 * t1 * dot(grad3[gi1], x1, y1, z1);
   }
 
-  double t2 = 0.6 - x2*x2 - y2*y2 - z2*z2;
+  float t2 = 0.6 - x2*x2 - y2*y2 - z2*z2;
   if(t2<0) n2 = 0.0;
   else {
       t2 *= t2;
       n2 = t2 * t2 * dot(grad3[gi2], x2, y2, z2);
   }
 
-  double t3 = 0.6 - x3*x3 - y3*y3 - z3*z3;
+  float t3 = 0.6 - x3*x3 - y3*y3 - z3*z3;
   if(t3<0) n3 = 0.0;
   else {
       t3 *= t3;
@@ -321,30 +321,30 @@ double noise( double x, double y, double z ) {
 
 
 // 4D raw Simplex noise
-double noise(double x, double y, double z, double w) {
+float noise(float x, float y, float z, float w) {
   // The skewing and unskewing factors are hairy again for the 4D case
-  double F4 = (sqrtf(5.0) - 1.0) / 4.0;
-  double G4 = (5.0 - sqrtf(5.0)) / 20.0;
-  double n0, n1, n2, n3, n4;  // Noise contributions from the five corners
+  float F4 = (sqrtf(5.0) - 1.0) / 4.0;
+  float G4 = (5.0 - sqrtf(5.0)) / 20.0;
+  float n0, n1, n2, n3, n4;  // Noise contributions from the five corners
 
   // Skew the (x,y,z,w) space to determine which cell of 24 simplices we're in
-  double s = (x + y + z + w) * F4;  // Factor for 4D skewing
+  float s = (x + y + z + w) * F4;  // Factor for 4D skewing
   int i = fastfloor(x + s);
   int j = fastfloor(y + s);
   int k = fastfloor(z + s);
   int l = fastfloor(w + s);
-  double t = (i + j + k + l) * G4;  // Factor for 4D unskewing
+  float t = (i + j + k + l) * G4;  // Factor for 4D unskewing
 
   // Unskew the cell origin back to (x,y,z,w) space
-  double X0 = i - t;  
-  double Y0 = j - t;
-  double Z0 = k - t;
-  double W0 = l - t;
+  float X0 = i - t;  
+  float Y0 = j - t;
+  float Z0 = k - t;
+  float W0 = l - t;
   // The x,y,z,w distances from the cell origin
-  double x0 = x - X0;
-  double y0 = y - Y0;
-  double z0 = z - Z0;
-  double w0 = w - W0;
+  float x0 = x - X0;
+  float y0 = y - Y0;
+  float z0 = z - Z0;
+  float w0 = w - W0;
 
   // For the 4D case, the simplex is a 4D shape I won't even try to describe.
   // To find out which of the 24 possible simplices we're in, we need to
@@ -389,22 +389,22 @@ double noise(double x, double y, double z, double w) {
   // The fifth corner has all coordinate offsets = 1, so no need to look that
   // up.
 
-  double x1 = x0 - i1 + G4;  // Offsets for second corner in (x,y,z,w) coords
-  double y1 = y0 - j1 + G4;
-  double z1 = z0 - k1 + G4;
-  double w1 = w0 - l1 + G4;
-  double x2 = x0 - i2 + 2.0 * G4;  // Offsets for third corner in (x,y,z,w) coords
-  double y2 = y0 - j2 + 2.0 * G4;
-  double z2 = z0 - k2 + 2.0 * G4;
-  double w2 = w0 - l2 + 2.0 * G4;
-  double x3 = x0 - i3 + 3.0 * G4;  // Offsets for fourth corner in (x,y,z,w) coords
-  double y3 = y0 - j3 + 3.0 * G4;
-  double z3 = z0 - k3 + 3.0 * G4;
-  double w3 = w0 - l3 + 3.0 * G4;
-  double x4 = x0 - 1.0 + 4.0 * G4;  // Offsets for last corner in (x,y,z,w) coords
-  double y4 = y0 - 1.0 + 4.0 * G4;
-  double z4 = z0 - 1.0 + 4.0 * G4;
-  double w4 = w0 - 1.0 + 4.0 * G4;
+  float x1 = x0 - i1 + G4;  // Offsets for second corner in (x,y,z,w) coords
+  float y1 = y0 - j1 + G4;
+  float z1 = z0 - k1 + G4;
+  float w1 = w0 - l1 + G4;
+  float x2 = x0 - i2 + 2.0 * G4;  // Offsets for third corner in (x,y,z,w) coords
+  float y2 = y0 - j2 + 2.0 * G4;
+  float z2 = z0 - k2 + 2.0 * G4;
+  float w2 = w0 - l2 + 2.0 * G4;
+  float x3 = x0 - i3 + 3.0 * G4;  // Offsets for fourth corner in (x,y,z,w) coords
+  float y3 = y0 - j3 + 3.0 * G4;
+  float z3 = z0 - k3 + 3.0 * G4;
+  float w3 = w0 - l3 + 3.0 * G4;
+  float x4 = x0 - 1.0 + 4.0 * G4;  // Offsets for last corner in (x,y,z,w) coords
+  float y4 = y0 - 1.0 + 4.0 * G4;
+  float z4 = z0 - 1.0 + 4.0 * G4;
+  float w4 = w0 - 1.0 + 4.0 * G4;
 
   // Work out the hashed gradient indices of the five simplex corners
   int ii = i & 255;
@@ -418,7 +418,7 @@ double noise(double x, double y, double z, double w) {
   int gi4 = perm[ii + 1 + perm[jj + 1 + perm[kk + 1 + perm[ll + 1]]]] % 32;
 
   // Calculate the contribution from the five corners
-  double t0 = 0.6 - x0 * x0 - y0 * y0 - z0 * z0 - w0 * w0;
+  float t0 = 0.6 - x0 * x0 - y0 * y0 - z0 * z0 - w0 * w0;
   if (t0 < 0)
     n0 = 0.0;
   else {
@@ -426,7 +426,7 @@ double noise(double x, double y, double z, double w) {
     n0 = t0 * t0 * dot(grad4[gi0], x0, y0, z0, w0);
   }
 
-  double t1 = 0.6 - x1 * x1 - y1 * y1 - z1 * z1 - w1 * w1;
+  float t1 = 0.6 - x1 * x1 - y1 * y1 - z1 * z1 - w1 * w1;
   if (t1 < 0)
     n1 = 0.0;
   else {
@@ -434,7 +434,7 @@ double noise(double x, double y, double z, double w) {
     n1 = t1 * t1 * dot(grad4[gi1], x1, y1, z1, w1);
   }
 
-  double t2 = 0.6 - x2 * x2 - y2 * y2 - z2 * z2 - w2 * w2;
+  float t2 = 0.6 - x2 * x2 - y2 * y2 - z2 * z2 - w2 * w2;
   if (t2 < 0)
     n2 = 0.0;
   else {
@@ -442,7 +442,7 @@ double noise(double x, double y, double z, double w) {
     n2 = t2 * t2 * dot(grad4[gi2], x2, y2, z2, w2);
   }
 
-  double t3 = 0.6 - x3 * x3 - y3 * y3 - z3 * z3 - w3 * w3;
+  float t3 = 0.6 - x3 * x3 - y3 * y3 - z3 * z3 - w3 * w3;
   if (t3 < 0)
     n3 = 0.0;
   else {
@@ -450,7 +450,7 @@ double noise(double x, double y, double z, double w) {
     n3 = t3 * t3 * dot(grad4[gi3], x3, y3, z3, w3);
   }
 
-  double t4 = 0.6 - x4 * x4 - y4 * y4 - z4 * z4 - w4 * w4;
+  float t4 = 0.6 - x4 * x4 - y4 * y4 - z4 * z4 - w4 * w4;
   if (t4 < 0) n4 = 0.0;
   else {
     t4 *= t4;
@@ -461,7 +461,7 @@ double noise(double x, double y, double z, double w) {
   return 27.0 * (n0 + n1 + n2 + n3 + n4);
 }
 
-double convertTo(double perlin, PerlinType type) {
+float convertTo(float perlin, PerlinType type) {
   switch (type) {
     case Layered:    return perlin * 0.5 + 0.5;     break;
     case Turbulence: return abs(perlin);            break;
