+++ src/objects/cylinder.h
@@ -5,10 +5,10 @@
 struct Cylinder : Object {
   Cylinder(BSDF *mat) : Object(mat) { bounds = AABB(1); };
 
-  bool hitBase(double Z, Ray &r, HitRec &rec);
+  bool hitBase(float Z, Ray &r, HitRec &rec);
   bool hitSide(Ray &r, HitRec &rec);
 
   bool hit(Ray& r, HitRec &rec);
-  Vec3 sample(double *pdf, RNG& rng);
+  Vec3 sample(float *pdf, RNG& rng);
 };
