+++ src/objects/surface_revolution.cc
@@ -3,6 +3,6 @@
 using namespace std;
 
 /* Define P(a, b) for ParametricSurface to use based on F */
-Vec3 SurfaceOfRevolution::P(double a, double b) {
+Vec3 SurfaceOfRevolution::P(float a, float b) {
   return RotateYMatrix(b) * F(a);
 }
